UpdatedDataStructure
Added more recent total cases and deaths. Put some order in the folder of the data set. Now every source of data is in a dif. folder. This will help create notes that describe the data and its origin. And of course, synced so that Politi and UN-Pop is now in the master files.